requirements.txt
amqp==2.5.0
anyjson==0.3.3
behave==1.2.6
billiard==3.6.0.0
celery==4.4.0rc2
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-behave==0.1.6
django-celery-results==1.1.2
django-debug-toolbar==2.0
django-flower==1.0.0
django-jinja==2.4.1
django-jinja2==0.1
gevent==1.4.0
greenlet==0.4.15
humanize==0.5.1
idna==2.8
Jinja2==2.10.1
kombu==4.6.3
MarkupSafe==1.1.1
parse==1.12.0
parse-type==0.5.2
plumbum==1.6.7
psycopg2-binary==2.8.3
pytz==2019.2
redis==3.3.6
requests==2.22.0
rpyc==4.1.1
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
whitenoise==4.1.3
